+++ backend/src/api/error/api.rs
@@ -3,11 +3,12 @@ use actix_web::http::header::ContentType;
 use actix_web::http::StatusCode;
 use actix_web::HttpResponse;
 use derive_more::{Display, Error};
-use log::error;
+use redis::RedisError;
 use sea_orm::DbErr;
 use serde::{Serialize, Serializer};
+use tracing::error;
 use utoipa::ToSchema;
-use validator::ValidationError;
+use validator::{ValidationError, ValidationErrors};
 
 use entity::error::EntityError;
 
@@ -50,15 +51,15 @@ impl ApiError {
     pub fn invalid_session() -> Self {
         Self {
             status_code: StatusCode::UNAUTHORIZED,
-            details: "Invalid session provided.".to_string(),
+            details: "Session expired or invalid!".to_string(),
             reference: None,
         }
     }
 
-    pub fn invalid_identity() -> Self {
+    pub fn session_limit_reached() -> Self {
         Self {
-            status_code: StatusCode::UNAUTHORIZED,
-            details: "Invalid identity provided.".to_string(),
+            status_code: StatusCode::CONFLICT,
+            details: "Session limit reached!".to_string(),
             reference: None,
         }
     }
@@ -102,6 +103,14 @@ impl ApiError {
             reference: SerializableStruct::new(&errors).ok(),
         }
     }
+
+    pub fn no_token_provided() -> Self {
+        Self {
+            status_code: StatusCode::UNAUTHORIZED,
+            details: "No token provided.".to_string(),
+            reference: None,
+        }
+    }
 }
 
 impl ResponseError for ApiError {
@@ -138,6 +147,16 @@ impl From<DbErr> for ApiError {
     }
 }
 
+impl From<RedisError> for ApiError {
+    fn from(value: RedisError) -> Self {
+        Self {
+            status_code: StatusCode::INTERNAL_SERVER_ERROR,
+            details: value.to_string(),
+            reference: None,
+        }
+    }
+}
+
 impl From<ValidationErrorJsonPayload> for ApiError {
     fn from(value: ValidationErrorJsonPayload) -> Self {
         let serializable_struct = SerializableStruct::new(&value).ok();
@@ -152,6 +171,12 @@ impl From<ValidationErrorJsonPayload> for ApiError {
 impl From<ValidationError> for ApiError {
     fn from(value: ValidationError) -> Self {
         Self::from(ValidationErrorJsonPayload::from(value))
+    }
+}
+
+impl From<ValidationErrors> for ApiError {
+    fn from(value: ValidationErrors) -> Self {
+        Self::from(ValidationErrorJsonPayload::from(&value))
     }
 }