+++ backend/src/api/mod.rs
@@ -1,7 +1,7 @@
-pub mod account;
-pub mod budget;
-pub mod currency;
 pub mod error;
+pub mod pagination;
+pub mod routes;
 pub mod status;
-pub mod transaction;
-pub mod user;
+
+// Use this once impl Trait is stable (https://github.com/rust-lang/rust/issues/63063)
+//pub type ApiResponse = Result<impl Responder, ApiError>;