+++ backend/src/api/pagination/mod.rs
@@ -1,0 +1,160 @@
+use actix_web::dev::Payload;
+use actix_web::error::QueryPayloadError;
+use actix_web::http::Uri;
+use actix_web::web::Query;
+use actix_web::{FromRequest, HttpRequest};
+use futures_util::future::LocalBoxFuture;
+use serde::{Deserialize, Serialize};
+use utoipa::openapi::path::{Parameter, ParameterBuilder, ParameterIn};
+use utoipa::openapi::{KnownFormat, ObjectBuilder, Required, SchemaFormat, SchemaType};
+use utoipa::{IntoParams, ToSchema};
+use validator::Validate;
+
+use crate::api::error::api::ApiError;
+use crate::wrapper::account::Account;
+use crate::wrapper::budget::Budget;
+use crate::wrapper::currency::Currency;
+use crate::wrapper::transaction::Transaction;
+
+pub const DEFAULT_PAGE: u64 = 1;
+pub const DEFAULT_LIMIT: u64 = 50;
+pub const MAX_LIMIT: u64 = 500;
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
+#[aliases(
+PaginatedAccount = Pagination < Account >,
+PaginatedBudget = Pagination < Budget >,
+PaginatedCurrency = Pagination < Currency >,
+PaginatedTransaction = Pagination < Transaction >
+)]
+pub struct Pagination<T: Serialize + ToSchema<'static>> {
+    #[serde(rename = "_metadata")]
+    pub metadata: Metadata,
+    pub data: Vec<T>,
+}
+
+impl<T: Serialize + ToSchema<'static>> Pagination<T> {
+    pub fn new(data: Vec<T>, page_size_param: &PageSizeParam, total: u64, uri: Uri) -> Self {
+        Self {
+            metadata: Metadata {
+                page: page_size_param.page,
+                limit: page_size_param.limit,
+                total,
+                links: Links::new(uri, page_size_param, total),
+            },
+            data,
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
+pub struct Metadata {
+    pub page: u64,
+    pub limit: u64,
+    pub total: u64,
+    pub links: Links,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
+pub struct Links {
+    pub prev: Option<String>,
+    pub next: Option<String>,
+}
+
+impl Links {
+    pub fn new(uri: Uri, page_size_param: &PageSizeParam, total: u64) -> Self {
+        let page = page_size_param.page;
+        let limit = page_size_param.limit;
+
+        let prev = if page > 1 {
+            Some(format!("{}?page={}&limit={}", uri.path(), page - 1, limit))
+        } else {
+            None
+        };
+        let next = if total >= limit {
+            Some(format!("{}?page={}&limit={}", uri.path(), page + 1, limit))
+        } else {
+            None
+        };
+
+        Self {
+            prev,
+            next,
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Validate, Deserialize)]
+pub struct PageSizeParam {
+    #[validate(range(min = 1))]
+    pub page: u64,
+    #[validate(range(min = 1, max = "MAX_LIMIT"))]
+    pub limit: u64,
+}
+
+impl PageSizeParam {
+    pub fn new(page: u64, limit: u64) -> Self {
+        Self {
+            page,
+            limit,
+        }
+    }
+}
+
+impl Default for PageSizeParam {
+    fn default() -> Self {
+        Self {
+            page: DEFAULT_PAGE,
+            limit: DEFAULT_LIMIT,
+        }
+    }
+}
+
+impl FromRequest for PageSizeParam {
+    type Error = ApiError;
+    type Future = LocalBoxFuture<'static, Result<Self, Self::Error>>;
+
+    fn from_request(req: &HttpRequest, _payload: &mut Payload) -> Self::Future {
+        let page_size_params: Result<Query<Self>, QueryPayloadError> = Query::from_query(req.query_string());
+
+        let page_size_params = match page_size_params {
+            Ok(page_size_params) => page_size_params.into_inner(),
+            Err(_) => Self::default(),
+        };
+
+        Box::pin(async move {
+            page_size_params.validate().map_err(ApiError::from)?;
+
+            Ok(page_size_params)
+        })
+    }
+}
+
+impl IntoParams for PageSizeParam {
+    fn into_params(parameter_in_provider: impl Fn() -> Option<ParameterIn>) -> Vec<Parameter> {
+        vec![
+            ParameterBuilder::new()
+                .name("page")
+                .required(Required::False)
+                .parameter_in(parameter_in_provider().unwrap_or_default())
+                .description(Some(format!("The page number. Default: {}", DEFAULT_PAGE)))
+                .schema(Some(
+                    ObjectBuilder::new()
+                        .schema_type(SchemaType::Integer)
+                        .format(Some(SchemaFormat::KnownFormat(KnownFormat::Int64))),
+                ))
+                .build(),
+            ParameterBuilder::new()
+                .name("limit")
+                .required(Required::False)
+                .parameter_in(parameter_in_provider().unwrap_or_default())
+                .description(Some(format!("The number of items per page. Default: {}", DEFAULT_LIMIT)))
+                .schema(Some(
+                    ObjectBuilder::new()
+                        .schema_type(SchemaType::Integer)
+                        .format(Some(SchemaFormat::KnownFormat(KnownFormat::Int64))),
+                ))
+                .build(),
+        ]
+    }
+}