+++ backend/src/api/routes/budget/controller.rs
@@ -1,8 +1,10 @@
+use actix_web::http::Uri;
 use actix_web::web::Path;
 use actix_web::{delete, get, patch, post, web, HttpResponse, Responder};
 use actix_web_validator::Json;
 
 use crate::api::error::api::ApiError;
+use crate::api::pagination::{PageSizeParam, Pagination};
 use crate::util::utoipa::{InternalServerError, ResourceNotFound, Unauthorized, ValidationError};
 use crate::wrapper::budget::dto::BudgetDTO;
 use crate::wrapper::budget::Budget;
@@ -18,10 +20,34 @@ pub fn budget_controller(cfg: &mut web::ServiceConfig) {
 
 #[utoipa::path(get,
 responses(
+(status = 200, description = "Successfully retrieved the Budgets.", content_type = "application/json", body = PaginatedBudget),
+(status = 401, response = Unauthorized),
+(status = 500, response = InternalServerError)
+),
+params(PageSizeParam),
+security(
+("bearer_token" = [])
+),
+path = "/api/v1/budget/?page={page}&size={size}",
+tag = "Budget"
+)]
+#[get("")]
+pub async fn get_all(user: Phantom<User>, page_size: PageSizeParam, uri: Uri) -> Result<impl Responder, ApiError> {
+    let total = Budget::count_all_by_user(user.get_id()).await?;
+    let budgets = Budget::find_all_by_user_paginated(user.get_id(), &page_size).await?;
+
+    Ok(HttpResponse::Ok().json(Pagination::new(budgets, &page_size, total, uri)))
+}
+
+#[utoipa::path(get,
+responses(
 (status = 200, description = "Successfully retrieved the Budget.", content_type = "application/json", body = Budget),
 (status = 401, response = Unauthorized),
 (status = 404, response = ResourceNotFound),
 (status = 500, response = InternalServerError)
+),
+security(
+("bearer_token" = [])
 ),
 path = "/api/v1/budget/{budget_id}",
 tag = "Budget"
@@ -36,28 +62,15 @@ pub async fn get_one(user: Phantom<User>, budget_id: Path<i32>) -> Result<impl R
     Ok(HttpResponse::Ok().json(budget))
 }
 
-#[utoipa::path(get,
-responses(
-(status = 200, description = "Successfully retrieved the Budgets.", content_type = "application/json", body = Vec<Budget>),
-(status = 401, response = Unauthorized),
-(status = 500, response = InternalServerError)
-),
-path = "/api/v1/budget",
-tag = "Budget"
-)]
-#[get("")]
-pub async fn get_all(user: Phantom<User>) -> Result<impl Responder, ApiError> {
-    let budgets = Budget::find_all(user.get_id()).await?;
-
-    Ok(HttpResponse::Ok().json(budgets))
-}
-
 #[utoipa::path(post,
 responses(
 (status = 200, description = "Successfully created the Budget.", content_type = "application/json", body = Budget),
 (status = 400, response = ValidationError),
 (status = 401, response = Unauthorized),
 (status = 500, response = InternalServerError)
+),
+security(
+("bearer_token" = [])
 ),
 path = "/api/v1/budget",
 tag = "Budget"
@@ -75,6 +88,9 @@ responses(
 (status = 401, response = Unauthorized),
 (status = 404, response = ResourceNotFound),
 (status = 500, response = InternalServerError)
+),
+security(
+("bearer_token" = [])
 ),
 path = "/api/v1/budget/{budget_id}",
 tag = "Budget"
@@ -97,6 +113,9 @@ responses(
 (status = 401, response = Unauthorized),
 (status = 404, response = ResourceNotFound),
 (status = 500, response = InternalServerError)
+),
+security(
+("bearer_token" = [])
 ),
 path = "/api/v1/budget/{budget_id}",
 tag = "Budget"