+++ backend/src/api/status/dto.rs
@@ -1,10 +1,15 @@
 use serde::Serialize;
+use time::OffsetDateTime;
 use utoipa::ToSchema;
+
+use entity::utility::time::get_now;
 
 #[derive(Serialize, ToSchema)]
 pub struct HealthResponse {
     pub healthy: bool,
     pub api_version: u8,
+    #[serde(with = "time::serde::rfc3339")]
+    pub local_time_now: OffsetDateTime,
     pub details: Option<String>,
 }
 
@@ -13,6 +18,7 @@ impl HealthResponse {
         Self {
             healthy,
             api_version: 1,
+            local_time_now: get_now(),
             details,
         }
     }