+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Unreleased
 
-- Nothing yet.
+- dependency updates
+- improved logging
+- Added logging file with log rotation.
+- added default headers to /api endpoints => `Content-Type: application/json` `Accept: application/json`
+- added rate-limiting to /api endpoints
+- added cors configuration
+- Added pagination to `get_all` endpoints
+- moved to token-based session authentication
+- new session limit (defaults to 25)
+- new session endpoints and session management
+- session can now be named
 
 # Version 0.1.0 (26.02.2024)